requirements.txt
aiogram==2.19
aioredis~=2.0
environs==9.5.0
gspread~=4.0
gspread-asyncio==1.5.3
oauth2client~=4.1.3
google-api-python-client==2.37.0
google-auth-httplib2
google-auth-oauthlib==0.4.6
pydantic